requirements.txt
absl-py==2.2.2
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.9.5
aiohttp-cors==0.8.1
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
ansicolors==1.1.8
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-timeout==4.0.3
atpublic==4.1.0
attrs==25.3.0
babel==2.17.0
backports.tarfile==1.2.0
beatrix_jupyterlab @ file:///tmp/environments/base/pip/packages/beatrix_jupyterlab-2024.66.154055.tar.gz#sha256=257afe97f249769790e978994d6554655e0c2be69c5faa3924a06c73bddbcab3
beautifulsoup4==4.13.3
bidict==0.23.1
bigframes==0.22.0
bleach==6.2.0
blessed==1.20.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-tpu-client==0.10
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cupy-cuda12x==13.4.1
cycler==0.12.1
Cython==3.0.12
dacite==1.9.2
dataproc_jupyter_plugin==0.1.80
db-dtypes==1.4.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
dm-tree==0.1.9
docker==7.1.0
docstring_parser==0.16
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.115.12
fastjsonschema==2.21.1
fastrlock==0.8.3
ffmpeg==1.4
filelock==3.18.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
gcsfs==2025.3.2
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==1.34.1
google-api-python-client==1.8.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.88.0
google-cloud-artifact-registry==1.15.2
google-cloud-bigquery==3.25.0
google-cloud-bigquery-connection==1.18.2
google-cloud-bigquery-storage==2.30.0
google-cloud-core==2.4.3
google-cloud-datastore==1.15.5
google-cloud-functions==1.20.2
google-cloud-iam==2.18.3
google-cloud-jupyter-config==0.0.10
google-cloud-language==2.17.1
google-cloud-monitoring==2.27.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.14.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gpustat==1.0.0
greenlet==3.1.1
-e git+https://github.com/wu-jeffrey-j/video-masking.git@cb39794721a965e2140b01d6286a247dd39a056c#egg=groundingdino&subdirectory=grounding_dino
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.49.0rc1
gymnasium==1.0.0
h11==0.14.0
hf-xet==1.1.4
htmlmin==0.1.12
httplib2==0.22.0
httptools==0.6.4
huggingface-hub==0.33.0
humanize==4.12.2
hydra-core==1.3.2
ibis-framework==7.1.0
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.4.0
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==6.29.5
ipython==8.21.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.1.6
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-http-over-ws==0.0.8
jupyter-server-mathjax==0.2.6
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_fileid==0.9.3
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab_git==0.44.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
jupytext==1.17.0
kernels-mixer==0.0.15
keyring==25.6.0
keyrings.google-artifactregistry-auth==1.1.2
kfp==2.5.0
kfp-pipeline-spec==0.2.2
kfp-server-api==2.0.5
kiwisolver==1.4.8
kubernetes==26.1.0
linkify-it-py==2.0.3
llvmlite==0.41.1
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.17.1
mistune==3.1.3
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multimethod==1.12
multipledispatch==1.0.0
narwhals==1.34.1
nbclassic==1.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbdime==3.2.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==6.5.7
notebook_executor @ file:///tmp/environments/base/pip/packages/notebook_executor
notebook_shim==0.2.4
numba==0.58.1
numpy==1.25.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==11.495.46
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.11.0.86
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ormsgpack==1.7.0
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
papermill==2.6.0
parso==0.8.4
parsy==2.1
patsy==1.0.1
pendulum==3.0.0
pexpect==4.9.0
phik==0.12.4
pillow==11.1.0
pins==0.8.7
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
portalocker==3.2.0
prettytable==3.16.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==3.20.3
psutil==5.9.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.22
pydantic==1.10.21
pydata-google-auth==1.9.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyogrio==0.10.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
# Editable install with no version control (SAM-2==1.0)
-e /home/wu_jeffrey_j/Grounded-SAM-2
scikit-learn==1.6.1
scipy==1.11.4
seaborn==0.12.2
SecretStorage==3.3.3
Send2Trash==1.8.3
shapely==2.1.0
simpervisor==1.0.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlglot==19.9.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
streamlit==1.46.0
supervision==0.25.1
sympy==1.14.0
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==9.1.2
tensorboardX==2.6.2.2
terminado==0.18.1
textual==3.0.1
threadpoolctl==3.6.0
time-machine==2.16.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
toolz==0.12.1
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.1
typeguard==4.4.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
uri-template==1.3.0
uritemplate==3.0.1
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.30.0
visions==0.7.5
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xxhash==3.5.0
y-py==0.6.2
yapf==0.43.0
yarl==1.19.0
ydata-profiling==4.6.0
ypy-websocket==0.8.4
zipp==3.21.0
gdown
google-cloud-storage
